Removed unsafe -m flag from su in init script
Revert "Added -m flag to pass env variables to the daemon"

This reverts commit 624c18997556c4c4c277fe5c056195cd90f122d9.

The flag was removed because it is potentially unsafe.

Per the BSD man page of su(1), if the target user's shell is not a standard login shell (which may apply for a "system user" as which pdagent runs), su will fail.