Switching back to using a custom HTTPS handler.
This allows us to pass `source_address` through to `HTTPSConnection`,
which otherwise isn't exposed at a high enough level for it to be useful
to us.

At some point we should also evaluate how necessary supporting
`source_address` is as it requires punching holes in abstractions that
we'd otherwise avoid.