Updates to testing and `six`-referencing.
This aims to fix some of the issues we had when referencing `six` under
some `pdagentd.py` execution cases where it was having trouble finding
it on the `PYTHONPATH`, plus is consistent with our `thirdparty`
conventions.

I've also done a drive-by fix of various things related to unit tests --
adjusting references to the new folder, adding `make test`, tuning some
of the logging settings to be less noisy under Python 3.